// apps/tui/src/themes.gen.ts
export interface Palette {
  background: string;
  backgroundPanel: string;
  backgroundElement: string;
  border: string;
  borderActive: string;
  borderSubtle: string;
  text: string;
  textMuted: string;
  primary: string;
  secondary: string;
  accent: string;
  error: string;
  warning: string;
  info: string;
  yellow: string;
  success: string;
  green: string;
}